requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-htmx==1.16.0
whitenoise==6.5.0
django-crispy-forms==2.0
crispy-bulma==0.8.0
PyJWT==2.8.0
python-telegram-bot==20.6
